requirements.txt
blinker==1.9.0
cachelib==0.13.0
click==8.3.0
colorama==0.4.6
Deprecated==1.2.18
Flask==3.1.2
Flask-Caching==2.3.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==4.0.0
flask-marshmallow==1.3.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
flasgger==0.9.7.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.6.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==4.0.1
marshmallow-sqlalchemy==1.4.2
mdurl==0.1.2
gunicorn==21.2.0
ordered-set==4.1.0
psycopg==3.1.18
packaging==25.0
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
rich==14.2.0
six==1.17.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
Werkzeug==3.1.3
wrapt==1.17.3
